[KYUUBI #4881] simplify code
