+++ externals/kyuubi-jdbc-engine/src/main/scala/org/apache/kyuubi/engine/jdbc/JdbcSQLEngine.scala
@@ -20,7 +20,7 @@ import org.apache.kyuubi.{Logging, Utils}
 import org.apache.kyuubi.Utils.{addShutdownHook, JDBC_ENGINE_SHUTDOWN_PRIORITY}
 import org.apache.kyuubi.config.KyuubiConf
 import org.apache.kyuubi.engine.jdbc.JdbcSQLEngine.currentEngine
-import org.apache.kyuubi.engine.jdbc.util.KyuubiJdbcUtil
+import org.apache.kyuubi.engine.jdbc.util.KyuubiJdbcUtils
 import org.apache.kyuubi.ha.HighAvailabilityConf.HA_ZK_CONN_RETRY_POLICY
 import org.apache.kyuubi.ha.client.RetryPolicies
 import org.apache.kyuubi.service.Serverable
@@ -73,7 +73,7 @@ object JdbcSQLEngine extends Logging {
 
       startEngine()
 
-      KyuubiJdbcUtil.initializeJdbcSession(kyuubiConf)
+      KyuubiJdbcUtils.initializeJdbcSession(kyuubiConf)
     } catch {
       case t: Throwable if currentEngine.isDefined =>
         currentEngine.foreach { engine =>
