+++ externals/kyuubi-jdbc-engine/src/main/scala/org/apache/kyuubi/engine/jdbc/dialect/DorisDialect.scala
@@ -163,8 +163,4 @@ class DorisDialect extends JdbcDialect {
     "doris"
   }
 
-  override def initializationSQLs(): Seq[String] = {
-    Seq("SHOW DATABASES")
-  }
-
 }
