+++ externals/kyuubi-jdbc-engine/src/main/scala/org/apache/kyuubi/engine/jdbc/session/JdbcSessionImpl.scala
@@ -25,6 +25,7 @@ import org.apache.hive.service.rpc.thrift.{TGetInfoType, TGetInfoValue, TProtoco
 import org.apache.kyuubi.KyuubiSQLException
 import org.apache.kyuubi.config.KyuubiReservedKeys.KYUUBI_SESSION_HANDLE_KEY
 import org.apache.kyuubi.engine.jdbc.connection.ConnectionProvider
+import org.apache.kyuubi.engine.jdbc.util.KyuubiJdbcUtils
 import org.apache.kyuubi.session.{AbstractSession, SessionHandle, SessionManager}
 
 class JdbcSessionImpl(
@@ -51,6 +52,7 @@ class JdbcSessionImpl(
       sessionConnection = ConnectionProvider.create(kyuubiConf)
       databaseMetaData = sessionConnection.getMetaData
     }
+    KyuubiJdbcUtils.initializeJdbcSession(kyuubiConf)
     super.open()
     info(s"The jdbc session is started.")
   }
